requirements.txt
click
Sphinx
coverage
python-dotenv>=0.5.1
numpy
matplotlib
seaborn
xlrd
pyyaml